Switched command line parsing to 'com.github.scopt:scopt' library
Instead of using custom, manual command line parsing I've switched
to 'com.github.scopt:scopt:3.3.0'. It makes parsing much easier
and we also get a nice, automatically-generated help page for free :)

On the downside, scopt internally invokes System.exit(0) when
you call it with -h/--help, so it became impossible to unit-test
that case - I had to remove these tests. In order to keep consistency,
I also call System.exit(0) manually when command line parsing failed,
as scopt displays a nice error message in that case and there is no need
to supplement it with anything. Which means another test had
to be removed...